+++ app/Console/Commands/Environment/EmailSettingsCommand.php
@@ -2,8 +2,8 @@
 
 namespace App\Console\Commands\Environment;
 
+use App\Traits\EnvironmentWriterTrait;
 use Illuminate\Console\Command;
-use App\Traits\Commands\EnvironmentWriterTrait;
 
 class EmailSettingsCommand extends Command
 {
