+++ app/Traits/EnvironmentWriterTrait.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace App\Traits\Commands;
+namespace App\Traits;
 
-use App\Exceptions\PanelException;
+use Exception;
 
 trait EnvironmentWriterTrait
 {
@@ -22,12 +22,13 @@ trait EnvironmentWriterTrait
 
     /**
      * Update the .env file for the application using the passed in values.
+     * @throws Exception
      */
     public function writeToEnvironment(array $values = []): void
     {
         $path = base_path('.env');
         if (!file_exists($path)) {
-            throw new PanelException('Cannot locate .env file, was this software installed correctly?');
+            throw new Exception('Cannot locate .env file, was this software installed correctly?');
         }
 
         $saveContents = file_get_contents($path);
